@@ -1,67 +1,66 @@
-﻿namespace Uceme.API.Controllers
+﻿namespace Uceme.API.Controllers;
+
+using System.Collections.Generic;
+using System.Data;
+using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using Uceme.Library.Services;
+using Uceme.Model.Models;
+
+[Authorize]
+[Route("api/[controller]")]
+[ApiController]
+public class HomeController : Controller
 {
-    using System.Collections.Generic;
-    using System.Data;
-    using System.Linq;
-    using Microsoft.AspNetCore.Authorization;
-    using Microsoft.AspNetCore.Mvc;
-    using Microsoft.Extensions.Logging;
-    using Uceme.Library.Services;
-    using Uceme.Model.Models;
+    private readonly ILogger<HomeController> logger;
 
-    [Authorize]
-    [Route("api/[controller]")]
-    [ApiController]
-    public class HomeController : Controller
+    private readonly IMedicoService medicoService;
+
+    private readonly IFotosService fotosService;
+
+    public HomeController(
+        IMedicoService medicoService,
+        IFotosService fotosService,
+        ILogger<HomeController> logger)
     {
-        private readonly ILogger<HomeController> logger;
+        this.medicoService = medicoService;
+        this.fotosService = fotosService;
+        this.logger = logger;
+    }
 
-        private readonly IMedicoService medicoService;
-
-        private readonly IFotosService fotosService;
-
-        public HomeController(
-            IMedicoService medicoService,
-            IFotosService fotosService,
-            ILogger<HomeController> logger)
+    [HttpGet("getmedicominvista")]
+    [AllowAnonymous]
+    public ActionResult<IEnumerable<Usuario>> GetMedicoMinVista()
+    {
+        IEnumerable<Usuario> result;
+        try
         {
-            this.medicoService = medicoService;
-            this.fotosService = fotosService;
-            this.logger = logger;
+            result = this.medicoService.GetMedicoMinVista(true);
+        }
+        catch (DataException)
+        {
+            return this.BadRequest();
         }
 
-        [HttpGet("getmedicominvista")]
-        [AllowAnonymous]
-        public ActionResult<IEnumerable<Usuario>> GetMedicoMinVista()
-        {
-            IEnumerable<Usuario> result;
-            try
-            {
-                result = this.medicoService.GetMedicoMinVista(true);
-            }
-            catch (DataException)
-            {
-                return this.BadRequest();
-            }
+        return result.ToList();
+    }
 
-            return result.ToList();
+    [HttpGet("mostrarfotos")]
+    [AllowAnonymous]
+    public ActionResult<IEnumerable<Fotos>> MostrarFotos()
+    {
+        IEnumerable<Fotos> listaFotos;
+        try
+        {
+            listaFotos = this.fotosService.GetFotos();
+        }
+        catch (DataException)
+        {
+            return this.BadRequest();
         }
 
-        [HttpGet("mostrarfotos")]
-        [AllowAnonymous]
-        public ActionResult<IEnumerable<Fotos>> MostrarFotos()
-        {
-            IEnumerable<Fotos> listaFotos;
-            try
-            {
-                listaFotos = this.fotosService.GetFotos();
-            }
-            catch (DataException)
-            {
-                return this.BadRequest();
-            }
-
-            return listaFotos.ToList();
-        }
+        return listaFotos.ToList();
     }
 }
