@@ -1,52 +1,51 @@
-﻿namespace Uceme.Foundation.Utilities
+﻿namespace Uceme.Foundation.Utilities;
+
+using System;
+using System.Net.Mail;
+using System.Threading.Tasks;
+
+public class SmtpClientWrapper : SmtpClient, ISmtpClient
 {
-    using System;
-    using System.Net.Mail;
-    using System.Threading.Tasks;
+    private bool disposed;
 
-    public class SmtpClientWrapper : SmtpClient, ISmtpClient
+    ~SmtpClientWrapper()
     {
-        private bool disposed;
+        this.Dispose(false);
+    }
 
-        ~SmtpClientWrapper()
+    public new string? Host { get; set; }
+
+    public new void Send(MailMessage mailMessage)
+    {
+        this.CheckDisposed();
+        base.Send(mailMessage);
+    }
+
+    public new async Task SendMailAsync(MailMessage mailMessage)
+    {
+        this.CheckDisposed();
+        await base.SendMailAsync(mailMessage).ConfigureAwait(false);
+    }
+
+    public new void Dispose()
+    {
+        this.Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    protected new virtual void Dispose(bool disposing)
+    {
+        if (!this.disposed)
         {
-            this.Dispose(false);
+            this.disposed = true;
         }
+    }
 
-        public new string? Host { get; set; }
-
-        public new void Send(MailMessage mailMessage)
+    protected void CheckDisposed()
+    {
+        if (this.disposed)
         {
-            this.CheckDisposed();
-            base.Send(mailMessage);
-        }
-
-        public new async Task SendMailAsync(MailMessage mailMessage)
-        {
-            this.CheckDisposed();
-            await base.SendMailAsync(mailMessage).ConfigureAwait(false);
-        }
-
-        public new void Dispose()
-        {
-            this.Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        protected new virtual void Dispose(bool disposing)
-        {
-            if (!this.disposed)
-            {
-                this.disposed = true;
-            }
-        }
-
-        protected void CheckDisposed()
-        {
-            if (this.disposed)
-            {
-                throw new ObjectDisposedException(nameof(SmtpClientWrapper));
-            }
+            throw new ObjectDisposedException(nameof(SmtpClientWrapper));
         }
     }
 }
