@@ -1,72 +1,71 @@
-﻿namespace Uceme.Library.Tests.Services
+﻿namespace Uceme.Library.Tests.Services;
+
+using System.Linq;
+using AutoMoqCore;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Uceme.Library.Services;
+using Uceme.Model.Data;
+
+[TestClass]
+public class AppointmentServiceTests
 {
-    using System.Linq;
-    using AutoMoqCore;
-    using Microsoft.EntityFrameworkCore;
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using Uceme.Library.Services;
-    using Uceme.Model.Data;
-
-    [TestClass]
-    public class AppointmentServiceTests
+    [TestMethod]
+    public void AppointmentService_RightInput_RightOutput()
     {
-        [TestMethod]
-        public void AppointmentService_RightInput_RightOutput()
+        //// ARRANGE
+        /////create In Memory Database
+        DbContextOptions<ApplicationDbContext> options = new DbContextOptionsBuilder<ApplicationDbContext>()
+            .UseInMemoryDatabase(databaseName: "EmployeeDataBase")
+            .Options;
+
+        //// Create mocked Context by seeding Data as per Schema///
+        using (ApplicationDbContext context = new ApplicationDbContext(options, new OperationalStoreOptionsMigrations()))
         {
-            //// ARRANGE
-            /////create In Memory Database
-            var options = new DbContextOptionsBuilder<ApplicationDbContext>()
-                .UseInMemoryDatabase(databaseName: "EmployeeDataBase")
-                .Options;
-
-            //// Create mocked Context by seeding Data as per Schema///
-            using (var context = new ApplicationDbContext(options, new OperationalStoreOptionsMigrations()))
+            context.Cita.Add(new Model.Models.Cita
             {
-                context.Cita.Add(new Model.Models.Cita
-                {
-                    dia = 1,
-                    email = "asd@as",
-                    hora = 1.1M,
-                    idCita = 1,
-                    idTurno = 1,
-                    nombre = "as",
-                    telefono = "123",
-                });
-                context.Cita.Add(new Model.Models.Cita
-                {
-                    dia = 1,
-                    email = "asd@as",
-                    hora = 1.1M,
-                    idCita = 2,
-                    idTurno = 1,
-                    nombre = "as",
-                    telefono = "123",
-                });
-                context.Turno.Add(new Model.Models.Turno
-                {
-                    idTurno = 1,
-                    idHospital = 1,
-                });
-                context.DatosProfesionales.Add(new Model.Models.DatosProfesionales
-                {
-                    idDatosPro = 1,
-                    nombre = "hospitalname",
-                });
+                dia = 1,
+                email = "asd@as",
+                hora = 1.1M,
+                idCita = 1,
+                idTurno = 1,
+                nombre = "as",
+                telefono = "123",
+            });
+            context.Cita.Add(new Model.Models.Cita
+            {
+                dia = 1,
+                email = "asd@as",
+                hora = 1.1M,
+                idCita = 2,
+                idTurno = 1,
+                nombre = "as",
+                telefono = "123",
+            });
+            context.Turno.Add(new Model.Models.Turno
+            {
+                idTurno = 1,
+                idHospital = 1,
+            });
+            context.DatosProfesionales.Add(new Model.Models.DatosProfesionales
+            {
+                idDatosPro = 1,
+                nombre = "hospitalname",
+            });
 
-                context.SaveChanges();
+            context.SaveChanges();
 
-                var mocker = new AutoMoqer();
-                mocker.SetInstance<IApplicationDbContext>(context);
+            AutoMoqer mocker = new AutoMoqer();
+            mocker.SetInstance<IApplicationDbContext>(context);
 
-                var sut = mocker.Create<AppointmentService>();
+            AppointmentService sut = mocker.Create<AppointmentService>();
 
-                //// ACT
-                var appointments = sut.GetAppointments();
+            //// ACT
+            System.Collections.Generic.IEnumerable<Model.Models.Appointment> appointments = sut.GetAppointments();
 
-                //// ASSERT
-                Assert.IsNotNull(appointments);
-                Assert.AreEqual(2, appointments.Count());
-            }
+            //// ASSERT
+            Assert.IsNotNull(appointments);
+            Assert.AreEqual(2, appointments.Count());
         }
     }
 }
