@@ -1,38 +1,37 @@
-﻿namespace Uceme.Library.Services
+﻿namespace Uceme.Library.Services;
+
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Linq;
+using Microsoft.Extensions.Logging;
+using Uceme.Model.Data;
+using Uceme.Model.Models;
+
+public class FotosService : IFotosService
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Data;
-    using System.Linq;
-    using Microsoft.Extensions.Logging;
-    using Uceme.Model.Data;
-    using Uceme.Model.Models;
+    private readonly ILogger<FotosService> logger;
 
-    public class FotosService : IFotosService
+    private readonly ApplicationDbContext context;
+
+    public FotosService(ILogger<FotosService> logger, IApplicationDbContext context)
     {
-        private readonly ILogger<FotosService> logger;
+        this.logger = logger;
+        this.context = (ApplicationDbContext)context;
+    }
 
-        private readonly ApplicationDbContext context;
-
-        public FotosService(ILogger<FotosService> logger, IApplicationDbContext context)
+    public IEnumerable<Fotos> GetFotos()
+    {
+        try
         {
-            this.logger = logger;
-            this.context = (ApplicationDbContext)context;
+            IQueryable<Fotos> listaFotos = this.context.Fotos.Where(o => o.destacada != null && o.destacada.Value);
+            this.logger.LogInformation($"retrieved {listaFotos.Count()} items");
+
+            return listaFotos;
         }
-
-        public IEnumerable<Fotos> GetFotos()
+        catch (Exception e)
         {
-            try
-            {
-                var listaFotos = this.context.Fotos.Where(o => o.destacada != null && o.destacada.Value);
-                this.logger.LogInformation($"retrieved {listaFotos.Count()} items");
-
-                return listaFotos;
-            }
-            catch (Exception e)
-            {
-                throw new DataException("Error retrieving fotos", e);
-            }
+            throw new DataException("Error retrieving fotos", e);
         }
     }
 }
