@@ -1,13 +1,12 @@
-﻿namespace Uceme.Library.Services
+﻿namespace Uceme.Library.Services;
+
+using System.Threading.Tasks;
+
+public interface IEmailService
 {
-    using System.Threading.Tasks;
+    Task<bool> SendEmailToManagementAsync(string fromAddress, string subject, string body);
 
-    public interface IEmailService
-    {
-        Task<bool> SendEmailToManagementAsync(string fromAddress, string subject, string body);
+    bool SendEmailToManagement(string fromAddress, string subject, string body);
 
-        bool SendEmailToManagement(string fromAddress, string subject, string body);
-
-        Task<bool> SendEmailToClientAsync(string toAddress, string subject, string body);
-    }
+    Task<bool> SendEmailToClientAsync(string toAddress, string subject, string body);
 }
