@@ -1,10 +1,9 @@
-﻿namespace Uceme.Library.Services
-{
-    using System.Collections.Generic;
-    using Uceme.Model.Models;
+﻿namespace Uceme.Library.Services;
 
-    public interface IFotosService
-    {
-        IEnumerable<Fotos> GetFotos();
-    }
+using System.Collections.Generic;
+using Uceme.Model.Models;
+
+public interface IFotosService
+{
+    IEnumerable<Fotos> GetFotos();
 }
