@@ -1,10 +1,9 @@
-﻿namespace Uceme.Library.Services
-{
-    using System.Collections.Generic;
-    using Uceme.Model.Models;
+﻿namespace Uceme.Library.Services;
 
-    public interface IMedicoService
-    {
-        IEnumerable<Usuario> GetMedicoMinVista(bool hackOrder);
-    }
+using System.Collections.Generic;
+using Uceme.Model.Models;
+
+public interface IMedicoService
+{
+    IEnumerable<Usuario> GetMedicoMinVista(bool hackOrder);
 }
