@@ -1,23 +1,22 @@
-﻿namespace Uceme.Library.Services
+﻿namespace Uceme.Library.Services;
+
+using System.Collections.Generic;
+using Uceme.Model.DataContracts;
+using Uceme.Model.Models;
+
+public interface ITechniqueService
 {
-    using System.Collections.Generic;
-    using Uceme.Model.DataContracts;
-    using Uceme.Model.Models;
+    Tecnica GetTechnique(int techniqueId);
 
-    public interface ITechniqueService
-    {
-        Tecnica GetTechnique(int techniqueId);
+    IEnumerable<Tecnica> GetTechniques();
 
-        IEnumerable<Tecnica> GetTechniques();
+    bool DeleteTechnique(int techId);
 
-        bool DeleteTechnique(int techId);
+    Tecnica UpdateTechine(Tecnica post);
 
-        Tecnica UpdateTechine(Tecnica post);
+    bool UpdateTechnique(TechniqueRequest postRequest);
 
-        bool UpdateTechnique(TechniqueRequest postRequest);
+    bool AddTechnique(TechniqueRequest postRequest);
 
-        bool AddTechnique(TechniqueRequest postRequest);
-
-        string GetNextTechImage();
-    }
+    string GetNextTechImage();
 }
