@@ -1,162 +1,187 @@
-﻿namespace Uceme.UI
+﻿namespace Uceme.UI;
+
+using System.IO;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.SpaServices.ReactDevelopmentServer;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
+using Uceme.Foundation.Utilities;
+using Uceme.Library.Services;
+using Uceme.Model.Data;
+using Uceme.Model.Models;
+using Uceme.Model.Settings;
+
+public class Startup
 {
-    using Microsoft.AspNetCore.Authentication;
-    using Microsoft.AspNetCore.Builder;
-    using Microsoft.AspNetCore.Hosting;
-    using Microsoft.AspNetCore.SpaServices.ReactDevelopmentServer;
-    using Microsoft.EntityFrameworkCore;
-    using Microsoft.Extensions.Configuration;
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Hosting;
-    using Microsoft.OpenApi.Models;
-    using Uceme.Foundation.Utilities;
-    using Uceme.Library.Services;
-    using Uceme.Model.Data;
-    using Uceme.Model.Models;
-    using Uceme.Model.Settings;
+    private readonly string relaxedPolicy = "RelaxedCorsPolicy";
 
-    public class Startup
+    private readonly string strictPolicy = "StrictCorsPolicy";
+
+    public Startup(IConfiguration configuration)
     {
-        private readonly string relaxedPolicy = "RelaxedCorsPolicy";
+        this.Configuration = configuration;
+    }
 
-        private readonly string strictPolicy = "StrictCorsPolicy";
+    public IConfiguration Configuration { get; }
 
-        public Startup(IConfiguration configuration)
+    // This method gets called by the runtime. Use this method to add services to the container.
+    public void ConfigureServices(IServiceCollection services)
+    {
+        IConfigurationSection appSettingsSection = this.Configuration.GetSection("AppSettings");
+        services.Configure<AuthMessageSenderSettings>(this.Configuration.GetSection("EmailSettings"));
+
+        services.Configure<AppSettings>(appSettingsSection);
+        string? ucemeConnection = this.Configuration.GetConnectionString("UcemeConnection");
+
+        if (ucemeConnection == null)
         {
-            this.Configuration = configuration;
+            throw new InvalidDataException("missing UcemeConnection settings");
         }
 
-        public IConfiguration Configuration { get; }
+        services.AddDbContext<ApplicationDbContext>(options =>
+            options.UseSqlServer(ucemeConnection));
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
+        SetupIdentity(services);
+        this.SetupCors(services);
+
+        services.AddSingleton<IConfiguration>(this.Configuration);
+
+        services.AddTransient<IEmailService, EmailService>();
+        services.AddTransient<ISmtpClient, SmtpClientWrapper>();
+        services.AddTransient<IEmailSender, EmailSender>();
+        services.AddTransient<IApplicationDbContext, ApplicationDbContext>();
+
+        // In production, the React files will be served from this directory
+        services.AddControllersWithViews();
+        services.AddRazorPages();
+        services.AddDatabaseDeveloperPageExceptionFilter();
+        services.AddSpaStaticFiles(configuration =>
         {
-            var appSettingsSection = this.Configuration.GetSection("AppSettings");
-            services.Configure<AuthMessageSenderSettings>(this.Configuration.GetSection("EmailSettings"));
+            configuration.RootPath = "ClientApp/build";
+        });
 
-            services.Configure<AppSettings>(appSettingsSection);
-
-            services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlServer(
-                    this.Configuration.GetConnectionString("UcemeConnection")));
-
-            SetupIdentity(services);
-            this.SetupCors(services);
-
-            services.AddSingleton<IConfiguration>(this.Configuration);
-
-            services.AddTransient<IEmailService, EmailService>();
-            services.AddTransient<ISmtpClient, SmtpClientWrapper>();
-            services.AddTransient<IEmailSender, EmailSender>();
-            services.AddTransient<IApplicationDbContext, ApplicationDbContext>();
-
-            // In production, the React files will be served from this directory
-            services.AddControllersWithViews();
-            services.AddRazorPages();
-            services.AddDatabaseDeveloperPageExceptionFilter();
-            services.AddSpaStaticFiles(configuration =>
-            {
-                configuration.RootPath = "ClientApp/build";
-            });
-
-            var swaggerSettings = this.Configuration.GetSection("SwaggerSettings").Get<SwaggerSettings>();
-            services.AddSwaggerGen(options =>
-            {
-                options.SwaggerDoc(swaggerSettings.SwaggerVersion, new OpenApiInfo { Title = swaggerSettings.SwaggerApp, Version = swaggerSettings.SwaggerVersion });
-            });
+        SwaggerSettings? swaggerSettings = this.Configuration.GetSection("SwaggerSettings").Get<SwaggerSettings>();
+        if (swaggerSettings == null)
+        {
+            throw new InvalidDataException("missing swaggerSettings settings");
         }
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        services.AddSwaggerGen(options =>
         {
+            options.SwaggerDoc(swaggerSettings?.SwaggerVersion, new OpenApiInfo { Title = swaggerSettings?.SwaggerApp, Version = swaggerSettings?.SwaggerVersion });
+        });
+    }
+
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
+        {
+            app.UseDeveloperExceptionPage();
+            app.UseDeveloperExceptionPage();
+            app.UseMigrationsEndPoint();
+
+            SwaggerSettings? swaggerSettings = this.Configuration.GetSection("SwaggerSettings").Get<SwaggerSettings>();
+            if (swaggerSettings == null)
+            {
+                throw new InvalidDataException("missing Swagger settings");
+            }
+
+            app.UseSwagger();
+            app.UseSwaggerUI(options =>
+            {
+                options.SwaggerEndpoint(swaggerSettings.SwaggerUri?.ToString(), swaggerSettings.SwaggerApp);
+            });
+        }
+        else
+        {
+            app.UseExceptionHandler("/Error");
+            //// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+            app.UseHsts();
+            app.UseHttpsRedirection();
+        }
+
+        CorsSettings? corsSettings = this.Configuration.GetSection("CorsSettings").Get<CorsSettings>();
+        if (corsSettings == null)
+        {
+            throw new InvalidDataException("missing corsSettings settings");
+        }
+
+        if (corsSettings.UseStrictPolicy)
+        {
+            _ = app.UseCors(this.strictPolicy);
+        }
+        else
+        {
+            _ = app.UseCors(this.relaxedPolicy);
+        }
+
+        app.UseStaticFiles();
+        app.UseSpaStaticFiles();
+
+        app.UseRouting();
+
+        app.UseAuthentication();
+        app.UseIdentityServer();
+        app.UseAuthorization();
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllerRoute(
+                name: "default",
+                pattern: "{controller}/{action=Index}/{id?}");
+            endpoints.MapRazorPages();
+        });
+
+        app.UseSpa(spa =>
+        {
+            spa.Options.SourcePath = "ClientApp";
+
             if (env.IsDevelopment())
             {
-                app.UseDeveloperExceptionPage();
-                app.UseDeveloperExceptionPage();
-                app.UseMigrationsEndPoint();
-
-                var swaggerSettings = this.Configuration.GetSection("SwaggerSettings").Get<SwaggerSettings>();
-                app.UseSwagger();
-                app.UseSwaggerUI(options =>
-                {
-                    options.SwaggerEndpoint(swaggerSettings.SwaggerUri?.ToString(), swaggerSettings.SwaggerApp);
-                });
+                spa.UseReactDevelopmentServer(npmScript: "start");
             }
-            else
-            {
-                app.UseExceptionHandler("/Error");
-                //// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-                app.UseHttpsRedirection();
-            }
+        });
+    }
 
-            var corsSettings = this.Configuration.GetSection("CorsSettings").Get<CorsSettings>();
-            if (corsSettings.UseStrictPolicy)
-            {
-                _ = app.UseCors(this.strictPolicy);
-            }
-            else
-            {
-                _ = app.UseCors(this.relaxedPolicy);
-            }
+    private static void SetupIdentity(IServiceCollection services)
+    {
+        services.AddDefaultIdentity<ApplicationUser>(options => options.SignIn.RequireConfirmedAccount = true)
+            .AddEntityFrameworkStores<ApplicationDbContext>();
 
-            app.UseStaticFiles();
-            app.UseSpaStaticFiles();
+        services.AddIdentityServer()
+            .AddApiAuthorization<ApplicationUser, ApplicationDbContext>();
 
-            app.UseRouting();
+        services.AddAuthentication()
+            .AddIdentityServerJwt();
+    }
 
-            app.UseAuthentication();
-            app.UseIdentityServer();
-            app.UseAuthorization();
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllerRoute(
-                    name: "default",
-                    pattern: "{controller}/{action=Index}/{id?}");
-                endpoints.MapRazorPages();
-            });
-
-            app.UseSpa(spa =>
-            {
-                spa.Options.SourcePath = "ClientApp";
-
-                if (env.IsDevelopment())
-                {
-                    spa.UseReactDevelopmentServer(npmScript: "start");
-                }
-            });
-        }
-
-        private static void SetupIdentity(IServiceCollection services)
+    private void SetupCors(IServiceCollection services)
+    {
+        CorsSettings? corsSettings = this.Configuration.GetSection("CorsSettings").Get<CorsSettings>();
+        if (corsSettings == null)
         {
-            services.AddDefaultIdentity<ApplicationUser>(options => options.SignIn.RequireConfirmedAccount = true)
-                .AddEntityFrameworkStores<ApplicationDbContext>();
-
-            services.AddIdentityServer()
-                .AddApiAuthorization<ApplicationUser, ApplicationDbContext>();
-
-            services.AddAuthentication()
-                .AddIdentityServerJwt();
+            throw new InvalidDataException("missing corsSettings settings");
         }
 
-        private void SetupCors(IServiceCollection services)
+        services.AddCors(o =>
         {
-            var corsSettings = this.Configuration.GetSection("CorsSettings").Get<CorsSettings>();
-            services.AddCors(o =>
+            o.AddPolicy(this.relaxedPolicy, builder =>
             {
-                o.AddPolicy(this.relaxedPolicy, builder =>
-                {
-                    builder.AllowAnyOrigin()
-                        .AllowAnyMethod()
-                        .AllowAnyHeader();
-                });
-
-                o.AddPolicy(this.strictPolicy, builder =>
-                {
-                    builder.WithOrigins(corsSettings.StrictPolicyHost ?? string.Empty)
-                            .WithMethods("PUT", "DELETE", "GET", "POST");
-                });
+                builder.AllowAnyOrigin()
+                    .AllowAnyMethod()
+                    .AllowAnyHeader();
             });
-        }
+
+            o.AddPolicy(this.strictPolicy, builder =>
+            {
+                builder.WithOrigins(corsSettings.StrictPolicyHost ?? string.Empty)
+                        .WithMethods("PUT", "DELETE", "GET", "POST");
+            });
+        });
     }
 }
