@@ -4,8 +4,8 @@ import parse from 'html-react-parser';
 import { Container } from 'reactstrap';
 import './Blogs.scss';
 import { Link } from 'react-router-dom';
-import tinyDate from '../../resources/images/tinydate.png';
-import photoIcon from '../../resources/images/photoicon.png';
+import tinyDate from '../../resources/images/tinydate.webp';
+import photoIcon from '../../resources/images/photoicon.webp';
 import BlogItem from '../../library/BlogItem';
 import SettingsContext from '../../SettingsContext';
 import BlogPostResponse from '../../library/BlogPostResponse';
