@@ -1,11 +1,11 @@
 /* eslint-disable react/function-component-definition */
 import * as React from 'react';
 import './Specialities.scss';
-import slide1 from '../../resources/images/cirugia-tiroidea.png';
-import slide2 from '../../resources/images/cirugia-paratiroidea.png';
-import slide3 from '../../resources/images/obesidad-morbida.png';
-import slide4 from '../../resources/images/cirugia-glandulas-suprarrenales.png';
-import slide5 from '../../resources/images/tiroidectomia-sin-cicatriz.png';
+import slide1 from '../../resources/images/cirugia-tiroidea.webp';
+import slide2 from '../../resources/images/cirugia-paratiroidea.webp';
+import slide3 from '../../resources/images/obesidad-morbida.webp';
+import slide4 from '../../resources/images/cirugia-glandulas-suprarrenales.webp';
+import slide5 from '../../resources/images/tiroidectomia-sin-cicatriz.webp';
 
 const specials = [
   {
